Stabalize IR pipeline and ZP preservation

+++ include/AST.hpp
@@ -31,12 +31,12 @@ class Statement : public ASTNode {};
 
 class IntegerLiteral : public Expression {
 public:
-    int value;
+    int64_t value;
     // Optional type from folded CastExpression (empty = default "int")
     std::string castType;
     int castPointerLevel = 0;
     bool castIsSigned = false;
-    IntegerLiteral(int v) : value(v) {}
+    IntegerLiteral(int64_t v) : value(v) {}
     void accept(ASTVisitor& visitor) override;
 };
